Remove init.rs and move resource_manager module
