+++ exonum-java-binding/core/rust/src/resource_manager/imp.rs
