+++ exonum-java-binding/core/rust/src/utils/mod.rs
@@ -22,7 +22,6 @@ mod jni;
 pub mod jni_cache;
 mod pair_iter;
 mod path;
-mod resource_manager;
 pub mod services;
 mod time_service;
 
@@ -36,4 +35,3 @@ pub use self::handle::{as_handle, cast_handle, drop_handle, to_handle, Handle};
 pub use self::jni::{get_class_name, get_exception_message};
 pub use self::pair_iter::PairIter;
 pub use self::path::executable_directory;
-pub use self::resource_manager::known_handles;
